+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] 2024-04-07
+### Added
+- Add `draft` scope and `post-status` support [#21](https://github.com/gRegorLove/indiebookclub/issues/21)
+- Show an error if a Micropub user did not grant `create` scope [#21]
+
+### Changed
+- Persist /new query parameters [#22](https://github.com/gRegorLove/indiebookclub/issues/22)
+
 ## [0.1.1] - 2023-12-02
 ### Added
 - Year in Review page for indiebookclub as a whole, e.g. `/review/2023`